pyproject.toml
[project]
name = "tcc_previsao_espectadores"
version = "1.0.0"
description = "\"Alteração do Trabalho de Conclusão de Curso para prever o número de espectadores de um filme\""
authors = [
    {name = "Joao Pedro Santos",email = "jpedro.csantos@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyarrow (>=19.0.0,<20.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "category-encoders (>=2.8.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
